refactor: Parser 리펙터링(테스트 복잡성으로 인한 클래스 분리)
