+++ src/main/kotlin/stringCalculator/domain/FormulaFormatException.kt
@@ -3,7 +3,10 @@ package stringCalculator.domain
 @Suppress("JavaIoSerializableObjectMustHaveReadResolve")
 sealed class FormulaFormatException : Throwable() {
     object CustomSeparatorFormatException : FormulaFormatException()
-    object NumberFormatException : FormulaFormatException()
+
+    object NotNumberFormatException : FormulaFormatException()
+
     object WrongFormatException : FormulaFormatException()
+
     object NegativeNumberException : FormulaFormatException()
 }
